Reworked whole thing into a static class
After using the original version of the tool it was a pain to work with
multiple items e.g. you have a column with a header and wanted to make
the header it's own block you would have to initialize two classes to
make it work now you just have to have two stacks in the one class.